harness: Be clearer about failed plans
When the plan of a deterministic test fails, that is to say, the number of
tests that were run don't match the number of specified tests, the harness
previously failed, as it should, but it did so much too silently. It only
marked the test bundle as failed, but didn't give a reason. This means, that
you would get output like "123 of 123 tests passed." with the bundle still
failing.

Now the harness marks the bundle with an oddity and lists all oddities in
summary, which makes the summary look like this:

    Processed 1 test bundle:
      • 1 of 1 bundle failed.
      • 1 of 1 bundle encountered an oddity.
      • test: (fail (plan (expected 21) (actual 20)))

    Processed 20 tests:
      • 20 of 20 tests passed.
      • 0 of 20 tests failed.

    Test Result: FAIL

Which is a bit better than before.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>