Move tap-harness program code to module
When using multiple Guile installations, it's easy to screw up the correct
guile binary to use. Guix solves this by rewriting the shebang line of the
script itself. On other systems this doesn't work. But if PATH is setup
properly, it is possible to just use the default guile binary.

This implements a shell spring board, that uses the default guile binary for
any given execution environment.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>