harness: Only search PATH if an executor program is specified
This is mostly for "prove" compatibility. With that harness you can do:

  prove -v -c 0001-test-some-stuff.t

With tap-harness before this, the same call structure fails:

  % tap-harness -v -c 0001-test-some-stuff.t
  Backtrace:
  …
     504:27  2 (program->state "0001-test-some-stuff.t" #<<harness-co…> …)
     480:15  1 (run-test _ _)
  In unknown file:
             0 (spawn "0001-test-some-stuff.t" ("0001-test-some-stuf…") …)
  ERROR: In procedure spawn:
  In procedure spawn: No such file or directory

With this change, the behaviour matches.

If a runner is specified, "r" is that runner's name. This needs to be looked up
in PATH. The test-suite code itself does not. So if no runner is specified, we
need to set #:search-path? to #f. Interestingly, you can't just use a truthy
value here. It must be a real boolean, in the #t case as well. Otherwise,
"spawn" greets you with:

  ERROR: In procedure spawn:
  Wrong type: "sh"

Hence, the "not not".

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>